# fully quantify native function invocation