--- a/src/PhpSpec/Formatter/Presenter/Differ/StringEngine.php
+++ b/src/PhpSpec/Formatter/Presenter/Differ/StringEngine.php
@@ -17,7 +17,7 @@ class StringEngine implements DifferEngine
 {
     public function supports($expected, $actual): bool
     {
-        return is_string($expected) && is_string($actual);
+        return \is_string($expected) && \is_string($actual);
     }
 
     public function compare($expected, $actual): string