--- a/src/PhpSpec/Formatter/Presenter/Exception/GenericPhpSpecExceptionPresenter.php
+++ b/src/PhpSpec/Formatter/Presenter/Exception/GenericPhpSpecExceptionPresenter.php
@@ -39,7 +39,7 @@ final class GenericPhpSpecExceptionPresenter extends AbstractPhpSpecExceptionPre
     {
         $lines  = explode(PHP_EOL, file_get_contents($file));
         $offset = max(0, $lineno - ceil($context / 2));
-        $lines  = array_slice($lines, $offset, $context);
+        $lines  = \array_slice($lines, $offset, $context);
 
         $text = PHP_EOL;
         foreach ($lines as $line) {