--- a/src/PhpSpec/Formatter/Presenter/Value/CallableTypePresenter.php
+++ b/src/PhpSpec/Formatter/Presenter/Value/CallableTypePresenter.php
@@ -45,8 +45,8 @@ final class CallableTypePresenter implements TypePresenter
      */
     public function present($value): string
     {
-        if (is_array($value)) {
-            $type = is_object($value[0]) ? $this->presenter->presentValue($value[0]) : $value[0];
+        if (\is_array($value)) {
+            $type = \is_object($value[0]) ? $this->presenter->presentValue($value[0]) : $value[0];
             return sprintf('%s::%s()', $type, $value[1]);
         }
 
@@ -54,8 +54,8 @@ final class CallableTypePresenter implements TypePresenter
             return '[closure]';
         }
 
-        if (is_object($value)) {
-            return sprintf('[obj:%s]', get_class($value));
+        if (\is_object($value)) {
+            return sprintf('[obj:%s]', \get_class($value));
         }
 
         return sprintf('[%s()]', $value);