--- a/src/PhpSpec/Formatter/Presenter/Value/QuotingStringTypePresenter.php
+++ b/src/PhpSpec/Formatter/Presenter/Value/QuotingStringTypePresenter.php
@@ -21,7 +21,7 @@ final class QuotingStringTypePresenter implements StringTypePresenter
      */
     public function supports($value): bool
     {
-        return 'string' === strtolower(gettype($value));
+        return 'string' === strtolower(\gettype($value));
     }
 
     /**