--- a/src/PhpSpec/Formatter/PrettyFormatter.php
+++ b/src/PhpSpec/Formatter/PrettyFormatter.php
@@ -66,7 +66,7 @@ final class PrettyFormatter extends ConsoleFormatter
             'broken' => $this->getStatisticsCollector()->getBrokenEvents(),
             'skipped' => $this->getStatisticsCollector()->getSkippedEvents(),
         ) as $status => $events) {
-            if (!count($events)) {
+            if (!\count($events)) {
                 continue;
             }
 
@@ -91,7 +91,7 @@ final class PrettyFormatter extends ConsoleFormatter
         }
 
         $io->write(sprintf("%d examples ", $this->getStatisticsCollector()->getEventsCount()));
-        if (count($counts)) {
+        if (\count($counts)) {
             $io->write(sprintf("(%s)", implode(', ', $counts)));
         }