--- a/src/PhpSpec/Formatter/ProgressFormatter.php
+++ b/src/PhpSpec/Formatter/ProgressFormatter.php
@@ -55,7 +55,7 @@ final class ProgressFormatter extends ConsoleFormatter
         $count = $stats->getEventsCount();
         $plural = $count !== 1 ? 's' : '';
         $io->write(sprintf("%d example%s ", $count, $plural));
-        if (count($counts)) {
+        if (\count($counts)) {
             $io->write(sprintf("(%s)", implode(', ', $counts)));
         }
 
@@ -120,7 +120,7 @@ final class ProgressFormatter extends ConsoleFormatter
             $size = $size - $length;
 
             if ($isDecorated) {
-                if ($length > strlen($text) + 2) {
+                if ($length > \strlen($text) + 2) {
                     $text = str_pad($text, $length, ' ', STR_PAD_BOTH);
                 } else {
                     $text = str_pad('', $length, ' ');
@@ -150,7 +150,7 @@ final class ProgressFormatter extends ConsoleFormatter
     {
         if ($io->isDecorated()) {
             $progressBar = implode('', $progress);
-            $pad = $this->getIO()->getBlockWidth() - strlen(strip_tags($progressBar));
+            $pad = $this->getIO()->getBlockWidth() - \strlen(strip_tags($progressBar));
             $io->writeTemp($progressBar.str_repeat(' ', $pad + 1).$total);
         } else {
             $io->writeTemp('/'.implode('/', $progress).'/  '.$total.' examples');