… src/PhpSpec/Listener/MethodReturnedNullListener.php
… src/PhpSpec/Listener/StatisticsCollector.php
… src/PhpSpec/Loader/StreamWrapper.php
… src/PhpSpec/Locator/PSR0/PSR0Locator.php
… src/PhpSpec/Matcher/ApproximatelyMatcher.php
… src/PhpSpec/Matcher/ArrayContainMatcher.php
… src/PhpSpec/Matcher/ArrayCountMatcher.php
… src/PhpSpec/Matcher/ArrayKeyMatcher.php
… src/PhpSpec/Matcher/ArrayKeyValueMatcher.php
… src/PhpSpec/Matcher/CallbackMatcher.php
… src/PhpSpec/Matcher/ComparisonMatcher.php
… src/PhpSpec/Matcher/IdentityMatcher.php
… src/PhpSpec/Matcher/Iterate/IterablesMatcher.php
… src/PhpSpec/Matcher/IterateAsMatcher.php
… src/PhpSpec/Matcher/ObjectStateMatcher.php
… src/PhpSpec/Matcher/ScalarMatcher.php
… src/PhpSpec/Matcher/StartIteratingAsMatcher.php
… src/PhpSpec/Matcher/StringContainMatcher.php
… src/PhpSpec/Matcher/StringEndMatcher.php
… src/PhpSpec/Matcher/StringRegexMatcher.php
… src/PhpSpec/Matcher/StringStartMatcher.php
… src/PhpSpec/Matcher/ThrowMatcher.php
… src/PhpSpec/Matcher/TraversableContainMatcher.php
… src/PhpSpec/Matcher/TraversableCountMatcher.php
… src/PhpSpec/Matcher/TraversableKeyMatcher.php
… src/PhpSpec/Matcher/TraversableKeyValueMatcher.php
… src/PhpSpec/Matcher/TriggerMatcher.php
… src/PhpSpec/Matcher/TypeMatcher.php
… src/PhpSpec/NamespaceProvider/ComposerPsrNamespaceProvider.php
… src/PhpSpec/ObjectBehavior.php