This is the 2nd go at https://github.com/phpspec/phpspec/pull/1148

I have prefixed all native function calls with `\` as suggested by @nicolas-grekas

Again, on the https://github.com/Sylius/Sylius testsuite, (just the code analysis stage)
Blackfire before:
https://blackfire.io/profiles/394774de-ade8-4c03-a62b-be9ab8060b75/graph
After:
https://blackfire.io/profiles/e2ee0855-f82a-4f3b-aaa6-10c78278b421/graph
Comparison:
https://blackfire.io/profiles/compare/7886df20-3b18-4070-a634-552013797fae/graph

Without blackfire:
before:
```
time bin/phpspec run
bin/phpspec run  6.04s user 0.25s system 99% cpu 6.311 total
```
after:
```
time bin/phpspec run
bin/phpspec run  5.40s user 0.22s system 99% cpu 5.636 total
```

Around 10% faster in real terms.