@nicolas-grekas 2017-09-12
There are only a few function where this is useful, you should not change all of them.
Here is the list:
https://github.com/php/php-src/blob/f2db305fa4e9bd7d04d567822687ec714aedcdb5/Zend/zend_compile.c#L3872