@bendavies 2017-09-12
@nicolas-grekas however, i assume there is no penalty for prefixing other methods?  The plus is that it is consistent in styling.