@avant1 2017-09-12
> i assume there is no penalty for prefixing other methods? The plus is that it is consistent in styling, so i am tempted to leave it.

From my point of view prefixing all function calls does not make much sense here. As I know it is not a common approach for modern PHP projects.
So it will be surprising to all new (and existing) contributors, and it will require us to add a code style check for that (not sure if Scrutinizer supports this) to make sure it will stay consistent in future.

Changing several most frequently called places with a proper comment will bring performance improvement without forcing changes to code style.