@bendavies 2017-09-12
@nicolas-grekas @avant1 now only prefixing the list of functions provided by nicolas.