@ciaranmcnulty 2017-09-12
Great stuff 👍

How can we add regular performance profiling? Is Blackfire free for open source?