@nicolas-grekas 2017-09-12
The Hack edition is free for everyone. For more advanced features, I don't think there is a plan for OSS.