@bendavies 2017-09-12
@ciaranmcnulty you could to what doctrine do and run https://github.com/phpbench/phpbench benchmarks as part of the build /cc @ocramius