@bendavies 2017-09-14
Thanks @ciaranmcnulty, and thanks @nicolas-grekas for the tips